Tasks_8/src/main/java/mordyasov_3_years_later/MyStringUtils.java
package mordyasov_3_years_later;

import java.util.Arrays;
import java.util.Map;
import java.util.function.Function;
import java.util.stream.Collectors;

public final class MyStringUtils {

    private MyStringUtils() {
    }

    public static String[] countingWords(String text, int n) {
        return Arrays.stream(text.split(" "))
                .map(String::toLowerCase)
                .collect(Collectors.groupingBy(Function.identity(), Collectors.summingInt(s -> 1)))
                .entrySet().stream()
                .sorted(Map.Entry.<String, Integer>comparingByValue().reversed().thenComparing(Map.Entry::getKey))
                .limit(n)
                .map(Map.Entry::getKey)
                .toArray(String[]::new);
    }
}
